Revert "farrah/fix: ads and chat"
