+++ components/order-chat.tsx
@@ -45,7 +45,7 @@ export default function OrderChat({
   const fileInputRef = useRef<HTMLInputElement>(null)
   const maxLength = 300
 
-  const { isConnected, getChatHistory, subscribe } = useWebSocketContext()
+  const { isConnected, getChatHistory, subscribe, leaveChannel } = useWebSocketContext()
 
   useEffect(() => {
     const unsubscribe = subscribe((data) => {
@@ -79,6 +79,12 @@ export default function OrderChat({
       }, 100)
     }
   }, [isConnected, getChatHistory, orderId])
+
+  useEffect(() => {
+    return () => {
+      leaveChannel("orders")
+    }
+  }, [])
 
   useEffect(() => {
     messagesEndRef.current?.scrollIntoView({ behavior: "smooth" })
